Refactoring Rest API app
