+++ lib/api/rest_api_base.rb
