+++ lib/generate_topology/operation_base.rb
