+++ lib/generate_topology/topology_generator.rb
